+++ src/RedisProxy.php
@@ -24,6 +24,8 @@ use RedisProxy\Driver\RedisDriver;
  * @method array hkeys(string $key) Get all fields in a hash (without values)
  * @method array hgetall(string $key) Get all fields and values in a hash
  * @method int hlen(string $key) Get the number of fields in a hash
+ * @method bool hexists(string $key, string $field) Check if <i>\$field</i> exists in hash <i>\$key</i>
+ * @method int hstrlen(string $key, string $field) Get the length of the string field, 0 if field does not exist
  * @method array smembers(string $key) Get all the members in a set
  * @method int scard(string $key) Get the number of members in a set
  * @method int sismember(string $key, string $member) Returns if member is a member of the set stored at key
@@ -40,6 +42,7 @@ use RedisProxy\Driver\RedisDriver;
  * @method array zpopmin(string $key, int $count = 1)
  * @method array zpopmax(string $key, int $count = 1)
  * @method array zrevrange(string $key, int $start, int $stop, bool $withscores = false) Return a range of members in a sorted set, by index, with scores ordered from high to low
+ * @method float zincrby(string $key, float $increment, string $member) Increment or decrement member of key by the given value (decrement when negative value is passed)
  * @method array publish(string $channel, string $message) Posts a message to the given channel
  */
 class RedisProxy
@@ -518,6 +521,20 @@ class RedisProxy
     {
         $this->init();
         return (float) $this->driver->call('hincrbyfloat', [$key, $field, $increment]);
+    }
+
+    /**
+     * Sets the expiration of one or more <i>\$fields</i> in given hash <i>\$key</i>, TTL is in <i>\$seconds</i>
+     * @param string|string[] ...$fields fields to set expiration on, either variadic enumeration or array of fields
+     * @return null|int[] int results for each field (in order) or `null` if no-key passed, return values are:
+     * -2 if field does not exist, 1 if expiration is set/updated and 2 if `HEXPIRE` is called with 0 seconds
+     * @throws RedisProxyException
+     */
+    public function hexpire(string $key, int $seconds, ...$fields): ?array
+    {
+        $fields = $this->prepareArguments('hexpire', ...$fields);
+        $this->init();
+        return $this->driver->call('hexpire', [$key, $seconds, ...$fields]);
     }
 
     /**